requirements.txt
requests==2.31.0
Flask==3.0.0
pandas==2.1.1
plotly==5.17.0
python-dotenv==1.0.0
